+++ index.js
@@ -2,7 +2,6 @@
 import { SnapcallListener } from './SnapcallListener'
 import { SnapcallParameter } from './SnapcallParameter'
 import { Snapcall } from './Snapcall'
-import React from 'react';